refactor : 마커인터페이스 대신 최상위 VO 클래스를 반환 타입으로 지정 (#800)
