+++ src/main/java/com/clubber/ClubberServer/domain/faq/controller/FaqController.java
@@ -2,7 +2,7 @@ package com.clubber.ClubberServer.domain.faq.controller;
 
 import com.clubber.ClubberServer.domain.faq.service.FaqService;
 import com.clubber.ClubberServer.global.config.swagger.DisableSwaggerSecurity;
-import com.clubber.ClubberServer.global.mapper.enums.EnumMapperVO;
+import com.clubber.ClubberServer.global.vo.enums.EnumMapperVO;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
