+++ src/main/java/com/clubber/ClubberServer/domain/faq/service/FaqService.java
@@ -1,7 +1,7 @@
 package com.clubber.ClubberServer.domain.faq.service;
 
 import com.clubber.ClubberServer.global.mapper.enums.EnumMapper;
-import com.clubber.ClubberServer.global.mapper.enums.EnumMapperVO;
+import com.clubber.ClubberServer.global.vo.enums.EnumMapperVO;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
