+++ src/main/java/com/clubber/ClubberServer/global/mapper/enums/EnumDefaultMapperType.java
@@ -1,11 +1,11 @@
 package com.clubber.ClubberServer.global.mapper.enums;
 
-import com.clubber.ClubberServer.global.vo.enums.EnumDefaultMapperVO;
+import com.clubber.ClubberServer.global.vo.enums.EnumMapperVO;
 
 public interface EnumDefaultMapperType extends EnumMapperType {
 
     @Override
     default EnumMapperVO createVO() {
-        return new EnumDefaultMapperVO(this);
+        return new EnumMapperVO(this);
     }
 }
