+++ src/main/java/com/clubber/ClubberServer/global/mapper/enums/EnumFaqMapperType.java
@@ -1,6 +1,7 @@
 package com.clubber.ClubberServer.global.mapper.enums;
 
 import com.clubber.ClubberServer.global.vo.enums.EnumFaQMapperVO;
+import com.clubber.ClubberServer.global.vo.enums.EnumMapperVO;
 
 public interface EnumFaqMapperType extends EnumMapperType {
