+++ src/main/java/com/clubber/ClubberServer/global/mapper/enums/EnumImageMapperType.java
@@ -1,6 +1,7 @@
 package com.clubber.ClubberServer.global.mapper.enums;
 
 import com.clubber.ClubberServer.global.vo.enums.EnumImageMapperVO;
+import com.clubber.ClubberServer.global.vo.enums.EnumMapperVO;
 
 public interface EnumImageMapperType extends EnumMapperType {
