+++ src/main/java/com/clubber/ClubberServer/global/mapper/enums/EnumMapperType.java
@@ -1,5 +1,7 @@
 package com.clubber.ClubberServer.global.mapper.enums;
 
+import com.clubber.ClubberServer.global.vo.enums.EnumMapperVO;
+
 public interface EnumMapperType {
 
 	String getCode();
