--- src/main/java/com/clubber/ClubberServer/global/mapper/enums/EnumMapperVO.java
@@ -1,4 +1,0 @@
-package com.clubber.ClubberServer.global.mapper.enums;
-
-public interface EnumMapperVO {
-}
