+++ src/main/java/com/clubber/ClubberServer/global/vo/enums/EnumFaQMapperVO.java
@@ -1,11 +1,10 @@
 package com.clubber.ClubberServer.global.vo.enums;
 
 import com.clubber.ClubberServer.global.mapper.enums.EnumFaqMapperType;
-import com.clubber.ClubberServer.global.mapper.enums.EnumMapperVO;
 import lombok.Getter;
 
 @Getter
-public class EnumFaQMapperVO extends EnumDefaultMapperVO implements EnumMapperVO {
+public class EnumFaQMapperVO extends EnumMapperVO {
 
 	private final String answer;
