+++ src/main/java/com/clubber/ClubberServer/global/vo/enums/EnumImageMapperVO.java
@@ -1,11 +1,10 @@
 package com.clubber.ClubberServer.global.vo.enums;
 
 import com.clubber.ClubberServer.global.mapper.enums.EnumImageMapperType;
-import com.clubber.ClubberServer.global.mapper.enums.EnumMapperVO;
 import lombok.Getter;
 
 @Getter
-public class EnumImageMapperVO extends EnumDefaultMapperVO implements EnumMapperVO {
+public class EnumImageMapperVO extends EnumMapperVO {
 
 	private final String imageUrl;
